Added base58 Address generation and Pk recovery
The Wallet class now has the address property.
This property is determined after the master keys are generated. We take the x point and the value of y mod 2 and base58 encode it for the address. Given the address, we can recover the (x,y) point.
There's a unit test to verify this.