Implemented VLI in UTXO - refactored
Refactored UTXO. We allow for a VLI to describe the locking script length (in the output) and the signature length (in the input).
The length of these strings will be given in BYTES. The strings themselves will be HEX chars. The size constants of the class will be given in BITS.