Added Block hex encoding and decoding
The Block will be internally divided into header and transactions. The number of transactions will be given by a VLI - the byte size of the header is fixed.

We also have a timestamp function which retrieves the epoch-seconds time. There's a decoder for these epoch-seconds as well.

Finally, added testing to unit test for utc conversion and block encoding/decoding.