Added basic mining functionality
Added mine_block functionality. Will take in raw_block and spit out updated raw_block.
Added increase_nonce function to block - the nonce is saved as hex_string, so we gave Block the ability to increase the nonce by one. This is called repeatedly by the miner.
The target is limited to 4 bytes = 32 bits. Hence, we take a target of the form
2 ** (256 - target_bits)
Which will give a number using up to 256 bits, which will be suitable for mining.